requirements.txt
Django>=4.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=24.2
channels>=4.3.2
channels-redis>=4.3.0
Pillow>=10.0.0